# docs/swagger.yaml
openapi: 3.0.0
info:
  title: traQ API
  description: traQ v2 API
  version: 2.6.3
  license:
    name: MIT
    url: 'https://github.com/traPtitech/traQ/blob/master/LICENSE'
  contact:
    name: traP
    url: 'https://github.com/traPtitech/traQ'
servers:
  - url: https://q.trap.jp/api/1.0
    description: production environment
  - url: https://traq-dev.tokyotech.org/api/1.0
    description: staging environment
paths:
...
  /users/{userID}:
    parameters:
      - $ref: "#/components/parameters/userIdInPath"
    get:
      tags:
        - user
      operationId: getUser
      description: ユーザーの詳細を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: |+
            正常に取得できました。
            ユーザー情報を返します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: +|
            正常に取得できませんでした。
            指定したユーザーは存在しません。

  # PATCH /users/{userID}

  # /users/{userID}/status

  # /users/{userID}/password
...
  /channels/{channelID}:
    parameters:
      - $ref: "#/components/parameters/channelIdInPath"
    get:
      tags:
        - channel
      operationId: getChannel
      description: チャンネルの情報を返します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: |+
            正常に取得ができました。
            チャンネルを返します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          description: |+
            取得に失敗しました。
            指定したチャンネルは存在しません。
    patch:
      tags:
        - channel
      operationId: editChannel
      description: チャンネルの情報を変更します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchChannel"
      responses:
        "204":
          description: 正常に変更ができました。
        "400":
          description: 失敗しました。リクエスト内容が不正です。
        "403":
          description: 失敗しました。権限がありません。
        "404":
          description: |+
            変更に失敗しました。
            指定したチャンネルは存在しません。
    delete:
      tags:
        - channel
      operationId: deleteChannel
      description: チャンネルを削除します。
      responses:
        "204":
          description: 削除に成功しました。
        "404":
          description: +|
            削除に失敗しました。
            指定したチャンネルは既に存在しません。
...
  /channels/{channelID}/bots:
    parameters:
      - $ref: "#/components/parameters/channelIdInPath"
    get:
      tags:
        - bot
      operationId: getChannelBots
      description: チャンネルに参加しているBot一覧を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelBot"
        "404":
          description: 取得できませんでした。指定したチャンネルは存在しません。
    post:
      tags:
        - bot
      operationId: addChannelBot
      description: チャンネルにBotを参加させます。
      security:
        - traqOAuth2:
            - write
            - manage_bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostBotChannel"
      responses:
        "200":
          description: 正常にBotが参加しました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotIdObject"
        "400":
          description: 参加させることができませんでした。リクエスト内容が不正です。
        "404":
          description: 参加させることができませんでした。指定したチャンネルは存在しません。
...
  /channels/{channelID}/webrtc/state:
    parameters:
      - $ref: "#/components/parameters/channelIdInPath"
    get:
      tags:
        - webrtc
      operationId: getChannelWebRTCState
      description: チャンネルのWebRTC状態を取得します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebRTCChannelState"
        "404":
          description: 取得できませんでした。指定したチャンネルは存在しません。
...
  /bots:
    get:
      tags:
        - bot
      operationId: getBots
      description: 自分が作成したBotの一覧を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      tags:
        - bot
      operationId: createBots
      description: Botを作成します。
      security:
        - traqOAuth2:
            - manage_bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostBot"
      responses:
        "201":
          description: 正常に作成できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotDetail"
        "400":
          description: 作成できませんでした。リクエスト内容が不正です。
        "409":
          description: 作成できませんでした。既に名前が使われています。
        "403":
          description: 作成できませんでした。権限がありません。
  /bots/{botID}:
    parameters:
      - $ref: "#/components/parameters/botIdInPath"
    get:
      tags:
        - bot
      operationId: getBot
      description: Botを取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: 取得できませんでした。指定したBotは存在しません。
    patch:
      tags:
        - bot
      operationId: editBot
      description: Bot情報を変更します。
      security:
        - traqOAuth2:
            - manage_bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBot"
      responses:
        "204":
          description: 正常に変更できました。
        "400":
          description: 変更できませんでした。リクエスト内容が不正です。
        "403":
          description: 変更できませんでした。権限がありません。
        "404":
          description: 削除できませんでした。指定したBotは存在しません。
    delete:
      tags:
        - bot
      operationId: deleteBot
      description: Botを削除します。
      security:
        - traqOAuth2:
            - manage_bot
      responses:
        "204":
          description: 正常に削除できました。
        "403":
          description: 削除できませんでした。権限がありません。
        "404":
          description: 削除できませんでした。指定したBotは存在しません。
...
  /bots/{botID}/icon:
    parameters:
      - $ref: "#/components/parameters/botIdInPath"
    get:
      tags:
        - bot
      operationId: getBotIcon
      description: 指定したBotのアイコンを取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        "404":
          description: 取得できませんでした。指定したBotは存在しません。
    put:
      tags:
        - bot
      operationId: changeBotIcon
      description: 指定したBotのアイコンを変更します。
      security:
        - traqOAuth2:
            - manage_bot
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: string
                  format: binary
                  description: Botユーザーのアイコン(1MBまでのpng, jpeg, gif)
            encoding:
              file:
                contentType: image/png, image/jpeg, image/gif
      responses:
        "204":
          description: 正常に変更できました。
        "400":
          description: 変更できませんでした。リクエスト内容が不正です。
        "403":
          description: 変更できませんでした。権限がありません。
        "404":
          description: 修正できませんでした。指定したBotは存在しません。
  /bots/{botID}/state:
    parameters:
      - $ref: "#/components/parameters/botIdInPath"
    put:
      tags:
        - bot
      operationId: changeBotState
      description: Botの状態を変更します。
      security:
        - traqOAuth2:
            - manage_bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutBotState"
      responses:
        "204":
          description: 変更できました。
        "202":
          description: Botのアクティベーションを受理しました。PINGリクエストが通り次第有効になります。
        "400":
          description: リクエスト内容が不正です。
        "403":
          description: 権限がありません。
        "404":
          description: 指定したBotは存在しません。
  /bots/{botID}/reissue:
    parameters:
      - $ref: "#/components/parameters/botIdInPath"
    post:
      tags:
        - bot
      operationId: reissueBotTokens
      description: Botの各種トークンを再発行します。
      security:
        - traqOAuth2:
            - manage_bot
      responses:
        "200":
          description: 再発行しました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotTokens"
        "403":
          description: 権限がありません。
        "404":
          description: 指定したBotは存在しません。
...
  /webrtc/state:
    get:
      tags:
        - webrtc
      operationId: getWebRTCState
      description: 自身のWebRTCの状態を取得します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebRTCUserState"
    put:
      tags:
        - webrtc
      operationId: putWebRTCState
      description: 自身のWebRTCの状態を変更します。
      security:
        - traqOAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutWebRTCState"
      responses:
        "204":
          description: 正常に変更できました。
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: r_session
    traqOAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /api/1.0/oauth2/authorize
          tokenUrl: /api/1.0/oauth2/token
          scopes:
            read: 読み取りスコープ
            write: 書き込みスコープ
            manage_bot: bot関連読み書きスコープ
  parameters:
    channelIdInPath:
      name: channelID
      description: 操作の対象となるチャンネルのID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userIdInPath:
      name: userID
      description: 操作の対象となるユーザーID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    botIdInPath:
      name: botID
      description: 操作の対象となるBotのID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    UUIDs:
      type: array
      description: UUIDの配列
      items:
        type: string
        format: uuid
...
    Channel:
      type: object
      properties:
        channelId:
          type: string
          format: uuid
          description: チャンネルUUID
        name:
          type: string
          description: チャンネル名
          example: yamada
        member:
          $ref: "#/components/schemas/UUIDs"
        parent:
          type: string
          format: uuid
          description: 親の階層のチャンネルUUID
        topic:
          type: string
          description: チャンネルトピック
          example: yamadaチャンネルです！
        children:
          $ref: "#/components/schemas/UUIDs"
        visibility:
          type: boolean
          description: チャンネルの可視状態
        force:
          type: boolean
          description: 強制通知チャンネルか
        private:
          type: boolean
          description: プライベートチャンネルか
        dm:
          type: boolean
          description: ダイレクトメッセージチャンネルか
...
    PatchChannel:
      type: object
      properties:
        name:
          type: string
          description: チャンネル名
          example: gps
        visibility:
          type: boolean
          description: 変更後のvisibility
          example: true
        force:
          type: boolean
          description: 強制通知かどうか
...
    UserStatus:
      type: integer
      enum:
        - 0
        - 1
        - 2
      description: アカウントの状態 (0:停止,1:有効,2:一時停止)
...
    UserDetail:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: ユーザーUUID
        name:
          type: string
          description: traQ ID
        displayName:
          type: string
          description: 表示名
        iconFileId:
          type: string
          format: uuid
          description: アイコンファイルUUID
        bot:
          type: boolean
          description: BOTかどうか
        twitterId:
          type: string
          description: ツイッターID
        lastOnline:
          type: string
          format: date-time
          description: 最終オンライン日時(オンラインに１度もなってない場合はnull)
        isOnline:
          type: boolean
          description: 現在オンラインかどうか
        suspended:
          type: boolean
          description: アカウントが停止中かどうか
        accountStatus:
          $ref: "#/components/schemas/UserStatus"
        tagList:
          $ref: "#/components/schemas/TagList"
...
    Tag:
      type: object
      properties:
        tagId:
          type: string
          format: uuid
          description: タグUUID
        tag:
          type: string
          description: タグ文字列
        isLocked:
          type: boolean
          description: タグがロックされているかどうか
        createdAt:
          type: string
          format: date-time
          description: タグ付与日時
        updatedAt:
          type: string
          format: date-time
          description: タグ更新日時
    TagList:
      type: array
      items:
        $ref: "#/components/schemas/Tag"
...
    Bot:
      type: object
      properties:
        botId:
          type: string
          format: uuid
          description: BOT UUID
        botUserId:
          type: string
          format: uuid
          description: BOTユーザーUUID
        description:
          type: string
          description: BOT説明
        subscribeEvents:
          type: array
          items:
            type: string
          description: BOTが購読しているイベントの配列
        state:
          type: integer
          description: BOT状態
        creatorId:
          type: string
          format: uuid
          description: BOT作成者UUID
        createdAt:
          type: string
          format: date-time
          description: BOT作成日時
        updatedAt:
          type: string
          format: date-time
          description: BOT更新日時
    ChannelBot:
      type: object
      properties:
        botId:
          type: string
          format: uuid
          description: BotのID
        botUserId:
          type: string
          format: uuid
          description: BotのユーザーID
    PostBotChannel:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          description: ボットコード
    BotIdObject:
      type: object
      properties:
        botId:
          type: string
          format: uuid
          description: 参加したBotのID
    PostBot:
      type: object
      required:
        - name
        - displayName
        - description
        - postUrl
      properties:
        name:
          type: string
          maxLength: 16
          description: botユーザーID。自動的に先頭に"BOT_"が付与されます
        displayName:
          type: string
          maxLength: 32
          description: botユーザーの表示名
        description:
          type: string
          description: botの説明
        postUrl:
          type: string
          format: url
          description: botのPOSTエンドポイント
    PatchBot:
      type: object
      properties:
        displayName:
          type: string
          description: Bot表示名
        description:
          type: string
          description: Bot説明
        privileged:
          type: boolean
          description: privileged属性
        webhookUrl:
          type: string
          format: url
          description: BotのPOSTエンドポイント
        creatorId:
          type: string
          format: uuid
          description: 移譲先のユーザーUUID
    PutBotState:
      type: object
      required:
        - state
      properties:
        state:
          type: string
          description: activeまたはinactive
    BotTokens:
      type: object
      properties:
        verificationToken:
          type: string
          description: Verification Token
        accessToken:
          type: string
          description: BOTアクセストークン
        botCode:
          type: string
          description: BOTインストールコード
    BotDetail:
      type: object
      properties:
        botId:
          type: string
          format: uuid
          description: BOT UUID
        botUserId:
          type: string
          format: uuid
          description: BOTユーザーUUID
        description:
          type: string
          description: BOT説明
        subscribeEvents:
          type: array
          items:
            type: string
          description: BOTが購読しているイベントの配列
        state:
          type: integer
          description: BOT状態
        creatorId:
          type: string
          format: uuid
          description: BOT作成者UUID
        createdAt:
          type: string
          format: date-time
          description: BOT作成日時
        updatedAt:
          type: string
          format: date-time
          description: BOT更新日時
        verificationCode:
          type: string
          example: NT672c7bgpfwdiMTyDXoeKiudlMazu
          description: BOT認証コード
        accessToken:
          type: string
          example: NT672c7bgpfwdiMTyDXoeKiudlMazu
          description: BOTアクセストークン
        postUrl:
          type: string
          format: uri
          description: BOTサーバーエンドポイント
          example: https://example.com/bot
        privileged:
          type: boolean
          description: 権限昇格されているかどうか
        botCode:
          type: string
          example: NT672c7bgpfwdiMTyDXoeKiudlMazu
          description: BOTインストールコード
...
    WebRTCUserState:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: ユーザーUUID
        channelId:
          type: string
          format: uuid
          description: ユーザーの接続先チャンネルUUID(未接続の場合は00000000-0000-0000-0000-000000000000)
        state:
          type: array
          items:
            type: string
          description: ユーザーの状態の配列(未接続の場合は空配列)
    WebRTCChannelState:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/WebRTCUserState"
          description: 接続ユーザーの状態の配列
    PutWebRTCState:
      type: object
      properties:
        channelId:
          type: string
          format: uuid
          nullable: true
          description: ユーザーの接続先チャンネルUUID(切断はnullを指定)
        state:
          type: array
          items:
            type: string
          description: ユーザーの状態の配列(切断は空配列を指定)
...
security:
  - cookieAuth: []